Fix tiling divider touch region.
The following CL includes:

* Swapping width with height, as the initial naming for SplitScreen meant
that width meant height when the screen is viewed horizontally as the
default is vertically.

* Fix a bug in dividerLeft bounds that used centerY instead of centerX.

* Fixing a bug where divider height is used instead of handle height for
handle height.

* This CL also fixes a bug where hovering on the divider would flip the
cursor to arrows when we only want it on the divider.

As this is a UI change, all those changes were tested extensively on
device.

Flag: com.android.window.flags.enable_tile_resizing
Test: On device testing, unit tests can't test correct touch regions.
Bug: 424342661
Change-Id: Iece4912ba45c8ce2d9a33199f4d7fcb41f1f5d08