Execute Keyguard-going-away on the changed Display
Keyguard-going-away was forcibly started on the default display even if
the call came in for a secondary display.

This causes a sync group timeout if the secondary display is off when
keyguardGoingAway is called. We fix it by marking the secondary display
as ready immediately too.

Flag: com.android.window.flags.keyguard_remove_default_display_usage
Test: atest KeyguardLockedTests
Bug: 430227391
Change-Id: I2b67fd31b56d908d3fd7cc6a50e4a26e9354a6b6