Fix MediaProjection Flicker tests
Flicker tests which use StartMediaProjectionAppHelper were failing due
to not being able to find the Calculator app from the app drawer - the
issue semed to be that sometimes it was scrolling past it in the
scenario. Addeed a check to see if the app is on thes screen before
scrolling.

Bug: 423974470
Test: StartAppMediaProjectionWithMaxDesktopWindows
Test: StartAppMediaProjectionResizeAndDrag
Flag: TEST_ONLY
Change-Id: Ife29fdd38804d3024d24c750409576fa300db2d2