Rename camera compat freeform -> simulate requested orientation.
Force-rotate policy is planned to be phased-out in favor of simulate requested orientation policy - this policy will extend to fullscreen and split-screen modes on large-screen devices.

CAMERA_COMPAT_ modes would be very long with the treatment description, and are only used in simulate requested orientation treatment, so this change is shortening its description, as well as a few other fields.

Please see child CLs where AppCompatCameraSimReqOrientationPolicy is used in different windowing modes.

Flag: EXEMPT renaming
Test: N/A
Bug: 432032916
Change-Id: I4668e0dca2cbd753f1c6f476b195b49d25d5dbdc