Block PiP state advancement if not in PiP
PipDisplayChangeObserver helps PiP component
keep track of system-wide display changes, and
can directly request to advance to bounds-changing
PipTransitionState states.

These should be blocked by PipTransitionState if
we are not in PiP.

We are also tightening the filter on the observer's side;
so if say system crash occurs while we are in PiP and
a pinned window remains on top, we don't want the presence
of the pinned window to be the only filter to advance to new states.

This same logic could apply to display changes when enter PiP
is scheduled but not yet complete: e.g. during fixed rotation.

Bug: 432672577
Flag: EXEMPT bugfix
Test: atest PipTransitionState
Test: atest PipDisplayChangeObserverTest
Change-Id: I4bf92a38ac2a92e241931527d9d4ef77072c4413