Remove unused removeImeSurface method
This method was previously ([1]) used in DisplayImeController, but with
the changes from [2], is no longer needed, so it can be removed. With
this we can drop the async suffix from the other method
removeImeSurfaceFromWindowAsync. Similarly, this also drops the async
suffix from reportPerceptibleAsync, and updates the javadoc.

  [1]: I4910c2a06cc67b0470477b245fc1de54b75f10f9
  [2]: I2b04517af59b949d30929db577c288b7232cfe65

Flag: EXEMPT cleanup
Bug: 281029564
Test: atest CtsInputMethodTestCases
Change-Id: Ieb4aa8410cabdb856d9b01b2926b6be0249869ce