Move platform lint tooling to lint_checks repo
frameworks/base/tools/lint was originally intended for running lint
checks against frameworks/base. This has since evolved to running lint
as a general tool across AOSP.

Move everything that is not specific to the `frameworks/base` repo to
the new `tools/lint_checks` repo that is available on all branches.

The primary motiviation is building `AndroidGlobalLintChecker` from
source, eliminating the need to copy the .jar build artifact to
`prebuilts` when it is updated.

Sources are copied as-is from the `frameworks/base` repo to the
`tools/lint_checks` repo from the root of a checkout:

- `mv frameworks/base/tools/lint/common tools/lint_checks/`
- `mv frameworks/base/tools/lint/fix toosl/lint_checks/`
- `mv frameworks/base/tools/lint/global tools/lint_checks/`
- `mv frameworks/base/tools/lint/utils tools/lint_checks/`

Bug: 264451752
Test: TH
Flag: EXEMPT pure refactor, host tooling only
Change-Id: I77826a0d2e4dd9a0b513405909de33fa7c154df7