[AE] Abort activity launch for finishing cross-uid TF
There is a race condition if a cross-uid embedded activity launches
another activity and then the TF is finished. In this case, depending on
the timing, ActivityStarter might place the new activity in a wrong
TaskFragment. For activities launched by a cross-uid embedded activity,
we always expect the new activity to be launched into the same TF. If
the TF is finishing, the best thing we can do seems to be aborting the
activity launch, as if it was launched into the TF and then finished
together with the TF.

Bug: 427881773
Test: Manually tested on Settings; atest ActivityStarterTest TaskFragmentTest
Flag: com.android.window.flags.activity_embedding_abort_cross_uid_launch_in_finishing_task_fragment
Change-Id: I65addba032ca8708b1d6906485cfe4f9ec9b1377