Fix onOpChanged null package regression
Fix an issue in Id0857e2aa1eb9906478135bffe483191b1452674 where we no
longer pass a null packageName to onOpChanged(String, String) in the
case of global restrictions.

That change was not backwards compatible because Java AIDL backend
permits nulls across non-null interfaces (for backwards compat), and
there is no client-side checking. So, despite annotations, nulls were
actually propagating to callers of the API, who could rely on it to
listen for global restriction changes for an op.

Fix this on the client-side by re-translating an empty package name
(which is the new service behavior) to null specifically for the legacy
param method, since the new, multi-param is explicitly marked non-null.
Update the docs for that method as well.

Bug: 434993506
Flag: EXEMPT bugfix, requires soak/backport
Test: manual logging in SystemAppOpsHelper
Change-Id: I6a6a696490993c077b7a86dfdd031a072e7460a2