Store lockout end time as rate limiter state
Instead of computing the lockout end time on-the-fly,
update it as state when successes/failures are reported,
and compute it on the first retrieval for the LskfId
after a cold boot. There is no observable behavior change
outside of the SoftwareRateLimiter.

The state will be queried in the next CL to implement
getLockoutEndTime. It will also track external timeouts.

Bug: 322014085
Flag: EXEMPT PURE_REFACTOR
Test: atest FrameworksServicesTests:com.android.server.locksettings
Change-Id: I5df3c2da835062575ab15524d8078d1251404bf0