Show Lock button only when device is unlocked
The Lock Power Menu button on desktop is mistakenly displayed when
the Power Menu is started from the Quick Settings when already on
the lock screen. This case is apparently not covered by the
"mKeyguardShowing" check.

Add a more generic check for the device to be unlocked to display
the Lock button in the Power Menu.

Bug: 435162937
Test: atest GlobalActionsDialogLiteTest
Test: manual verification on desktop device
Flag: NONE. The global action will only be configured for desktop
Change-Id: I258261462450713574ae91fcc5f2ef2244e52312