[flexiglass] alwaysCompose QS scene
This is a performance optimization where we make content in the QS scene
(single shade) always compose (e.g. compose but
remain invisible, even when not current content). This helps with the
heavy cost of composing this content, which is burdened by a lot of
Text.

Note that this CL doesn't change the default minActiveState of the
rememberViewModel calls in either class. The default of STARTED is safe
because it will only be achieved when the transition to the content is
complete.

The CL also makes TileListener utilize LaunchedEffectWithLifecycle
instead of just LaunchedEffect so tiles don't listen forever.

Bug: 430488243
Fix: 435719827
Test: manually verified that I can navigate to the scene and that its interactive to touch
Test: turned on verbose logging for QSLog and filtered for "listening"
watched it switch from false to true when opening QQS and then QS and
from true back to false as I closed QS and collapsed the shade
Flag: com.android.systemui.scene_container

Change-Id: Ifccb459b47f3f4d1da334fe41bac427ba3e0c230