Organize SysUI display subcomponents/modules the same as the main ones
This includes naming convention and organization of modules and
component inheritance.

PerDisplayCommonModule becomes PerDisplaySystemUIModule. This is the
equivalent to SystemUIModule, and is included in all variants of
SystemUI.

PerDisplayPhoneModule becomes PerDisplayReferenceSystemUIModule. This
is the AOSP module that is the "default" implementation of SystemUI.

Test: Build & Run all SysUI variants
Bug: 417927751
Flag: EXEMPT Dagger changes that can't be flagged
Change-Id: Id2a9017adab9cbdef5cff4ae4f59f9d32689c6f7