Update KeyguardInteractor#primaryBouncerShowing to include showingSoon
When primaryBouncerShowingSoon is true, this means the bouncer
background is showing but the bouncer contents themselves (ie:
pin/password/pattern views) haven't shown yet. For the rest of
keyguard, this essentially is equivalent to the bouncer showing.

When the showing state doesn't include the showingSoon state,
there are race conditions where the KeyguardTransitionFramework
can get stuck in a stale PrimaryBouncer transition state
when the PrimaryBouncerInteractor/Repo state reflects the
primary bouncer isn't showing.

Test: atest KeygaurdInteractorTest
Fixes: 432471590
Flag: EXEMPT bugfix
Change-Id: I249a5ca1bd8d808ab05ed09323db2168f463010f