fix stale bundle header state after open QS
Occasionally, [ShadeInteractor.isShadeFullyCollapsed] will emit `true`
when QS is fully expanded, which was causing us to update the bundle
header to collapsed, even though the bundle itself (tracked in the
GroupExpansionManager) is still expanded.

After some digging, the [snapTo] used inside of [BundleInteractor] was
determined to be unnecessary, so by moving the [lastCollapseTime] update
to the code path where the bundle is collapsed when the shade is closed,
we can remove the usage of [isShadeFullyCollapsed] entirely.

Flag: com.android.systemui.notification_bundle_ui
Fixes: 435245683
Test: manual
Change-Id: Iaa75b5e0576579d08a172c9bcf8db03642c5af47