[WindowDecor] Trigger relayout when system theme changes.
This CL implements onThemeChanged for DesktopModeWindowDecorViewModel so
that we can trigger a relayout when the system color changes to update
the header color for tasks that are already open. A boolean is also
added to WindowDecoration so that the AppHeaderViewHolder will be
recreated with the new context -- this will update the cached color
scheme accordingly.

Recall: http://recall/clips/4d457d63-3c90-41d9-9118-55faf0eac95c

Bug: 412489198
Flag: EXEMPT bugfix
Test: atest WMShellUnitTests:com.android.wm.shell.windowdecor
Test: Manual - navigate to Wallpaper & Style in Settings > change the
Home screen color > observe that the header changes color along with the
app content

Change-Id: I1ebabd7e16d0fa4a71f889564781d28b1b303596