3.6 Implement a modern version of leash setup.
The new DefaultTransitionHelper implements the same leash setup logic
as RemoteAnimationRunneCompat. It is introduced as an interface so
that Keyguard can have its own custom implementation (upcoming).

Trying to make each of the CLs as small as possible to keep them
digestible and low risk. For the refactor plan see
go/animlib-shell-refactor-plan.

Bug: 397180418
Flag: com.android.systemui.animation_library_shell_migration
Test: atest ActivityTransitionAnimatorTest + manual
Change-Id: I9d0ecb50188e66be50619a3f34c5261f2a328581