Query KeyguardStateController for the previous state
In exitKeyguardAndFinishSurfaceBehindRemoteAnimation, we check the
previous showing state to decide whether we need to hide the keyguard
view after the remote animation.

With the new keygaurd transition, the mShowing in KVM is directly set to
the target state to make it in sync with the WM's final state. To check
the mShowing in the KVM doesn't provide the needed information. Instead,
check the value in KeyguardStateController which reflects the existance
of the keyguard view visibility to make it work.

Bug: 364930619
Test: MultiDisplayKeyguardTests
Flag: com.android.window.flags.ensure_keyguard_does_transition_starting_bug_fix
Change-Id: I59bde420d6e7f54942fc4d7dd5e15dcc5a46ce6c