Check for Adj display groups to fallback to state=off
When an external display is connected, and the internal display times
out first, when the external display finally times out, and the internal
goes to AOD, there is a flicker that occurs on the screen.

One of the causes of this, is that the fallback of display policies is
DIM. We End up going from OFF->DIM->OFF->DOZE. By falling back to OFF in
this particular circumstance, we end up only going from OFF->DOZE which
is correct behaviour to help resolve the cause of this flicker.

Changes are guarded by bugfix flag, and we also validate that we are on
the default display group, and that other adjacent displays exist and
are all non-interactive before exercising this new code branch.

Additionally, there is an animation that plays, we want to skip this if
the default display is non-interactive.

The combination of these two changes completely removes the flicker.

Bug: 425275163
Bug: 436832920
Flag: com.android.server.power.feature.flags.separate_timeouts_flicker
Test: atest PowerGroupTest
Change-Id: I085134ab577d882bbdd17a24b0c0701f9a54c990