Support excluding caption bar from task snapshot
This adds the registerExcludeLayersFromTaskSnapshot() method for
TaskOrganizer. It provides a way for clients to register a set of
layers that should be excluded when a task snapshot is taken,
specifically for layers not managed by the Window Manager.
Also extract the bgExecutor to WindowDecoration.

Bug: 399228526
Flag: com.android.window.flags.excluding_layer_from_task_snapshot
Test: enable desktop windowing mode, open app, move the task to back
then enter Overivew, verify the caption bar won't show in the
TaskView

Change-Id: I14f6c72499118e4f7308a1053f152d0a5f334be1