[flexi] Add LockscreenFrontScrim
LockscreenFrontScrim.kt replaces scrim_in_front which no longer
exists in flexiglass. This scrim is used to dim AOD UI to
simulate lower brightness levels on devices that don't support
lower nits while in the DOZE display state when the brightness
sensor reports low lux.

Do not include wallpaper dimming here (in the
future this will be added in flexiglass
as part of the LockscreenBackScrim which is
not yet implemented).

Bug: 433477971
Flag: com.android.systemui.scene_container
Test: atest DozeServiceHostTest
Test: atest DozeScreenBrightnessTest
Test: manual (enter AOD, observe scrim
on devices that use it)

Change-Id: Iec37e9735160d0901d0988c25966bc1447544e92