Support creating a filtered snapshot from an existing snapshot computer.
It is a bit awkward that, in some places in Package Manager, for calling
ART Service methods, we create a filtered snapshot from a new snapshot
computer while we already have a computer. It leads to insistency when
there's a race.

This CL reuses the existing snapshot computer to avoid the awkwardness.

Bug: 258223472
Test: Presubmit
Flag: android.content.pm.alternative_for_dexopt_cleanup
Change-Id: If9968708e5e32f83c3f043527cf77157d986ac58