Ignore app layer assertion of RestartAppInSizeCompatModeTest
The app layer is optional to keep visible because the app is killed
and the surface can be hidden or destroyed during relaunching.
And it may not be directly related to the appearance depending on
the implementation of transition (e.g. cover a snapshot), so ignore
it as a note.

Bug: 435398625
Flag: TEST_ONLY
Test: RestartAppInSizeCompatModeTest
Change-Id: Iee06d82ee4556520de7b3ee7f7647a1546f09a73