[Expressive] disable scrolling in landscape mode
In the new UI, the collapsing toolbar will be collapsed by default. To
maintain a consistent user experience, especially when multiple
fragments are sharing the same activity, we need to disable the
scrolling functionality of the collapsing toolbar when the device is in
landscape mode. This will prevent inconsistent behavior and ensure the
toolbar's state remains predictable.

Bug: 424864280
Test: visual verify
Flag: EXEMPT bugfix
Change-Id: I0a03cc3c3284a91eaf958446ec56de349688a42a