Large screen capture: adding settings menu
This CL adds the dropdown menu that is triggered by pressing the cog
button in the pre capture toolbar. The dropdown has four settings, three
of which are disabled when the capture type is SCREENSHOT. This CL adds
the data to the PreCaptureViewModel, but the settings are not yet hooked
up to any capture functionality.

Bug: 437971334
Test: ScreenCaptureRecordParametersModelInteractorTest, Manual
Flag: com.android.systemui.large_screen_screencapture

Change-Id: I61a7d097ac7f041f5870e5c09e8f41a7c61d7ea8