Avoid creating a new StatusBarRootFactory every time
The StatusBarRootFactory was not scoped as PerDisplaySingleton so we were creating a new instance every time it was requested from the component.

This in turns was creating a new instance of DarkIconManager (as a factory was used in StatusBarRootFactory), that was registering a dumpable multiple times, causing a crash.

Bug: 438213567
Test: Dagger bindings are not really testable
Flag: com.android.systemui.shared.status_bar_connected_displays
Change-Id: If2ec51cfcfa58208ef63adfb2062e7d03e940b1b